refactor: move course location util to common djangoapp instead of xmodule
